refactor: single task expanded view, new condensed task card
